requirements.txt
kubernetes==25.3.0
pyOpenSSL==22.1.0
